Rename FormStateTag component to FormStatusTagComponent
