--- app/components/form_state_tag_component/view.html.erb
@@ -1,1 +1,0 @@
-<%= govuk_tag(text: state, colour: status_colour) %>
