+++ app/components/form_status_tag_component/view.html.erb
@@ -1,0 +1,1 @@
+<%= govuk_tag(text: status, colour: status_colour) %>
