--- spec/components/form_state_tag_component/form_state_tag_component_preview.rb
@@ -1,9 +1,0 @@
-class FormStateTagComponent::FormStateTagComponentPreview < ViewComponent::Preview
-  def default
-    render(FormStateTagComponent::View.new)
-  end
-
-  def live_status
-    render(FormStateTagComponent::View.new(state: "live"))
-  end
-end
