+++ spec/components/form_status_tag_component/form_status_tag_component_preview.rb
@@ -1,0 +1,9 @@
+class FormStatusTagComponent::FormStatusTagComponentPreview < ViewComponent::Preview
+  def default
+    render(FormStatusTagComponent::View.new)
+  end
+
+  def live_status
+    render(FormStatusTagComponent::View.new(status: "live"))
+  end
+end
